# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProcessMiner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/t/{tenant_slug}/systems/{system_slug}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get System",
        "operationId": "get_system_api_t__tenant_slug__systems__system_slug__get",
        "parameters": [
          {
            "name": "system_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Slug"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/t/{tenant_slug}/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_t__tenant_slug__artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "records"
        ],
        "summary": "Delete Artifact",
        "operationId": "delete_artifact_api_t__tenant_slug__artifacts__artifact_id__delete",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/artifacts/{artifact_id}/match": {
      "post": {
        "tags": [
          "records"
        ],
        "summary": "Match Artifact",
        "description": "어느 양식의 기록인지 모델에게 (다시) 묻는다.",
        "operationId": "match_artifact_api_t__tenant_slug__artifacts__artifact_id__match_post",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/artifacts/{artifact_id}/template": {
      "put": {
        "tags": [
          "records"
        ],
        "summary": "Set Artifact Template",
        "description": "양식을 확정한다(모델의 제안을 받아들이거나 직접 고른다). 이어서 항목 값을 뽑는다.",
        "operationId": "set_artifact_template_api_t__tenant_slug__artifacts__artifact_id__template_put",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactTemplateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/systems/{system_slug}/records": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "List Records",
        "description": "발행한 기록. 최근 것이 먼저.",
        "operationId": "list_records_api_t__tenant_slug__systems__system_slug__records_get",
        "parameters": [
          {
            "name": "system_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Slug"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordOut"
                  },
                  "title": "Response List Records Api T  Tenant Slug  Systems  System Slug  Records Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/records/{record_id}": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Get Record",
        "operationId": "get_record_api_t__tenant_slug__records__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "records"
        ],
        "summary": "Update Record",
        "operationId": "update_record_api_t__tenant_slug__records__record_id__patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/records/{record_id}/publish": {
      "post": {
        "tags": [
          "records"
        ],
        "summary": "Publish Record",
        "operationId": "publish_record_api_t__tenant_slug__records__record_id__publish_post",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "performed_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed On"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "match_state": {
            "type": "string",
            "title": "Match State"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Confidence"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MatchCandidateOut"
            },
            "type": "array",
            "title": "Candidates"
          },
          "record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecordSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "kind",
          "size_bytes",
          "char_count",
          "title",
          "performed_on",
          "state",
          "match_state",
          "template",
          "match_confidence",
          "candidates",
          "record",
          "run",
          "created_at",
          "actions",
          "segments"
        ],
        "title": "ArtifactDetail"
      },
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "performed_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed On"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "match_state": {
            "type": "string",
            "title": "Match State"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Confidence"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MatchCandidateOut"
            },
            "type": "array",
            "title": "Candidates"
          },
          "record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecordSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "kind",
          "size_bytes",
          "char_count",
          "title",
          "performed_on",
          "state",
          "match_state",
          "template",
          "match_confidence",
          "candidates",
          "record",
          "run",
          "created_at",
          "actions"
        ],
        "title": "ArtifactOut"
      },
      "ArtifactTemplateIn": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "ArtifactTemplateIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecordPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "performed_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed On"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/RecordFieldIn"
            },
            "type": "array",
            "title": "Fields",
            "default": []
          },
          "confirm": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Confirm",
            "default": []
          }
        },
        "type": "object",
        "title": "RecordPatch"
      },
...
      "SystemDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "org_unit_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Unit Id"
          },
          "parent_system_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent System Id"
          },
          "root_system_id": {
            "type": "string",
            "format": "uuid",
            "title": "Root System Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "org_unit_id",
          "parent_system_id",
          "root_system_id",
          "created_at",
          "document_count",
          "actions"
        ],
        "title": "SystemDetail"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}